Creamos el singleton conexion

+++ Pj_ConexionMySQL/frmAlumnos.cs
@@ -1,4 +1,6 @@
 using MySql.Data.MySqlClient;
+using Repository;
+using System.Configuration;
 using System.Data;
 
 namespace Pj_ConexionMySQL
@@ -7,15 +9,15 @@ namespace Pj_ConexionMySQL
     public partial class frmAlumnos : Form
     {
         //variables de clase (globales)
-        public MySqlConnection conexion;
+        //public MySqlConnection conexion;
         DataTable tabla;
        
 
         public frmAlumnos()
         {
             InitializeComponent();
-            String cadenaConexion = "server=localhost;user=root;password=root;database=lenguaje1";
-            conexion = new MySqlConnection(cadenaConexion);
+            //String cadenaConexion = "server=localhost;user=root;password=root;database=lenguaje1";
+            //conexion = new MySqlConnection(cadenaConexion);
 
         }
 
@@ -28,8 +30,13 @@ namespace Pj_ConexionMySQL
         {
             try
             {
+                MySqlDataReader dataReader = null;
+                var sqlConn = new MySqlConnection();
+                sqlConn = Conexion.getInstancia("root", "root").CrearConexion();
+                sqlConn.Open();
+
                 String query = "SELECT * FROM alumnos";
-                MySqlCommand comando = new MySqlCommand(query, conexion);
+                MySqlCommand comando = new MySqlCommand(query, sqlConn);
                 MySqlDataAdapter adaptador = new MySqlDataAdapter(comando);
                 DataTable tabla = new DataTable();
                 adaptador.Fill(tabla);
@@ -49,8 +56,12 @@ namespace Pj_ConexionMySQL
         {
             try
             {
+                var sqlConn=new MySqlConnection();
+                sqlConn = Conexion.getInstancia("root", "root").CrearConexion();
+                sqlConn.Open();
+
                 String query = "select * from alumnos";
-                MySqlCommand comando = new MySqlCommand(query, conexion);
+                MySqlCommand comando = new MySqlCommand(query, sqlConn);
                 MySqlDataReader reader = comando.ExecuteReader();
                 tabla = new DataTable();
                 tabla.Load(reader);
@@ -106,7 +117,7 @@ namespace Pj_ConexionMySQL
                     int id = Convert.ToInt32(this.dataGridView1.CurrentRow.Cells["id"].Value);
                     MessageBox.Show($"nombre seleccionado: {nombre}");
 
-                    Form alumno = new frmAlumnosForm(conexion, id, nombre, dni,2,this);
+                    Form alumno = new frmAlumnosForm(id, nombre, dni,2,this);
                     alumno.Show();
                 }
                 else
@@ -129,18 +140,35 @@ namespace Pj_ConexionMySQL
             //int id = Convert.ToInt32(this.dataGridView1.CurrentRow.Cells["id"].Value);
             //MessageBox.Show($"nombre seleccionado: {nombre}");
 
-            Form alumno = new frmAlumnosForm(conexion, 1,this);
+            Form alumno = new frmAlumnosForm(1,this);
             alumno.Show();
         }
 
         private void button5_Click(object sender, EventArgs e)
         {
-            string id = this.dataGridView1.CurrentRow.Cells["id"].Value.ToString();
-            string query = "delete from alumnos where id=" + id;
-            MySqlCommand comando = new MySqlCommand(query, conexion);
-            int resul = comando.ExecuteNonQuery();
-            MessageBox.Show($"La consulta devolvio: {resul}");
+            eliminarAlumno(this.dataGridView1.CurrentRow.Cells["id"].Value.ToString());
             obtenerTodos();
+        }
+
+        private void eliminarAlumno(string id)
+        {
+            
+            DialogResult a = MessageBox.Show("Esta seguro que desea eliminar al alumno?","Sistema alumnado",MessageBoxButtons.YesNo,MessageBoxIcon.Question);
+            if(a == DialogResult.Yes)
+            {
+                using (var sqlConn = Conexion.getInstancia("root", "root").CrearConexion())
+                {
+                    sqlConn.Open();
+
+                    string query = "delete from alumnos where id=" + id;
+                    MySqlCommand comando = new MySqlCommand(query, sqlConn);
+                    int resul = comando.ExecuteNonQuery();
+                    MessageBox.Show($"La consulta devolvio: {resul}");
+                }
+                
+            }
+
+            
         }
 
         private void frmAlumnosLista_Load(object sender, EventArgs e)
@@ -148,7 +176,7 @@ namespace Pj_ConexionMySQL
          
             try
             {
-                conexion.Open();
+                //conexion.Open();
                 this.txtBuscar.Enabled = true;
                 obtenerTodos();//nada
                 //MessageBox.Show("Conexión exitosa");
@@ -164,23 +192,36 @@ namespace Pj_ConexionMySQL
         {
             try
             {
-                String query = "select * from alumnos";
-                MySqlCommand comando = new MySqlCommand(query, conexion);
-                MySqlDataReader reader = comando.ExecuteReader();
-                tabla = new DataTable();
-                tabla.Load(reader);
-                dataGridView1.DataSource = tabla;
-                //while (reader.Read())
-                //{
-                //    Console.WriteLine(reader["nombre"].ToString());
-                //    MessageBox.Show(reader["nombre"].ToString());   
-                //}
+                //var sqlConn = new MySqlConnection();
+                //using llama a Dispose() es decir cierra la conexión automaticamente
+                //no hace falta sqlConn.Close(); 
+                using (var sqlConn = Conexion.getInstancia("root", "root").CrearConexion())
+                {
+                    sqlConn.Open();
+
+                    String query = "select * from alumnos";
+                    MySqlCommand comando = new MySqlCommand(query, sqlConn);
+                    MySqlDataReader reader = comando.ExecuteReader();
+                    tabla = new DataTable();
+                    tabla.Load(reader);
+                    dataGridView1.DataSource = tabla;
+                    //while (reader.Read())
+                    //{
+                    //    Console.WriteLine(reader["nombre"].ToString());
+                    //    MessageBox.Show(reader["nombre"].ToString());   
+                    //}
+
+                }
 
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.Message);
                 MessageBox.Show("Revise el usuario y contraseña de la base de datos");
+            }
+            finally
+            {
+                
             }
         }
     }
